pyproject.toml
[tool.poetry]
name = "sufc-stats"
version = "0.1.0"
description = ""
authors = ["Chris Hunt <chris.hunt@hudl.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.23.4"
mplsoccer = "^1.1.8"
jupyterlab = "^3.5.0"
python-dotenv = "^0.21.0"
matplotlib = "^3.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
